+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added read partitions in parallel for topic listener.
+
 ## v3.109.0
 * Added control plane fields for split-merge topics (Create,Alter,Describe)